Refactor coloring utilities to a private `ridgeplot._coloring` package
