+++ src/ridgeplot/_coloring/css_colors.py
